feature(junit-per-account): junit handling when ACCOUNT_TO_ASSUME is set
Under the previous behavior when running against a loop of accounts, each
loop would delete the junit files from the previous loop.  If they did
not then the files would collide.

This change adds account_id to the junit xml filenames and skips the
deletion of previous junit files.  Only when ACCOUNT_TO_ASSUME (and
ROLE_TO_ASSUME) is set.

+ adds account_id to the output files.  TEST-$account_id-$num
+ does not delete previous junit files

I only use -A when I am running against multiple accounts in a loop
on a scheduled run in my CI/CD.  I want all of the junit results to show
and not clobber or delete each other.  I clean the junit directory once
at the start.